Frontend: Restructure frontend code
