+++ src/main/reactjs/.babelrc
