+++ src/main/reactjs/.gitignore
@@ -1,8 +1,8 @@
 .idea/
 .vscode/
-node_modules/
+**/node_modules/
 dist
-.DS_Store
+**/.DS_Store
 *.tgz
 lerna-debug.log
 npm-debug.log*
