+++ src/main/reactjs/package.json
