+++ src/main/reactjs/public/index.html
