+++ src/main/reactjs/src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import { App } from './app';
+import { App } from './App';
 
 ReactDOM.render(<App />, document.getElementById('root'));
