+++ src/main/reactjs/tsconfig.json
